requirements.txt
Django>=4.0,<5.0
psycopg2-binary>=2.9
whitenoise>=6.0
dj-database-url>=1.0
gunicorn>=20.0
python-dotenv>=0.21
djangorestframework>=3.13
django-cors-headers>=3.13
django-allauth>=0.54
dj-rest-auth>=2.2
drf-yasg>=1.21
